[native] Move static functions from PrestoTask to Util.
