+++ presto-native-execution/presto_cpp/main/TaskManager.cpp
@@ -1112,7 +1112,7 @@ void TaskManager::shutdown() {
     PRESTO_SHUTDOWN_LOG(INFO)
         << "Waited (" << seconds
         << " seconds so far) for 'Running' tasks to complete. " << numTasks
-        << " tasks left: " << PrestoTask::taskNumbersToString(taskNumbers);
+        << " tasks left: " << util::taskNumbersToString(taskNumbers);
     std::this_thread::sleep_for(std::chrono::seconds(1));
     taskNumbers = getTaskNumbers(numTasks);
     ++seconds;
